==> Step06_nomogram_dca.Rmd <==
---
title: "Nomogram analysis"
author: "Xiong Wang"
date: "2023-10-28"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```
# 1. Univariate Cox regression analysis

```{r}
rm(list=ls())
load("data/riskscore_model.Rdata")
library(tinyarray)

surv=risk_model
dat1 = surv_cox(t(surv[,3:10]),surv,continuous = T,pvalue_cutoff = 1)
dat2 = as.data.frame(dat1[,c("HR","HRCILL","HRCIUL","p")])

datp = format(round(dat2[,1:3],3),nsmall = 3)

dat2$Trait = rownames(dat1)
dat2$HR2 = paste0(datp[, 1], "(", datp[, 2], "-", datp[, 3], ")")
dat2$p = ifelse(
  dat2$p <= 0.001,
  "<0.001",
  ifelse(
    dat2$p <= 0.01,
    "<0.01",
    ifelse(
      dat2$p <= 0.05,
      paste(round(dat2$p, 3), "*"),
      round(dat2$p, 3)
    )
  )
)



# plot
labeltext=cbind(c("Factor",dat2$Trait),
                c("HR (95%CI)",dat2$HR2),
                c("P Value",dat2$p))
library(forestplot)
forestplot(
  labeltext,
  mean = c(NA,dat2[, 1]),
  lower = c(NA,dat2[, 2]),
  upper = c(NA,dat2[, 3]),
  zero = 1,
  boxsize = 0.2,
  title="Hazard Ratio Plot (Univariate Cox)",
  xlab="Odds Ratio",
  is.summary=c(TRUE,rep(FALSE,8)),
  col = fpColors(box = '#1075BB', lines = 'black', zero = 'grey'),
  lty.ci = "solid",
  graph.pos = 3,#箱线图位置
  txt_gp=fpTxtGp(label = gpar(cex=1.1),
                 ticks = gpar(cex=1.1),
                 xlab = gpar(cex=1.1),
                 title = gpar(cex=1.2))
)

```

# 2. Multiivariate Cox regression analysis

```{r}
rm(list=ls())
load("data/riskscore_model.Rdata")

library(survival)
library(survminer)

surv=risk_model
model1 = coxph(Surv(time,event)~age+N+T+grade+riskscore,data = surv)

m = summary(model1)

# HR and 95%CI
dat2 = as.data.frame(round(m$conf.int[, c(1, 3, 4)], 2))
dat2 = tibble::rownames_to_column(dat2, var = "Trait")
dat2$p = round(m$coefficients[, 5], 3)
colnames(dat2)[2:5] = c("HR", "lower", "upper","p")

# HR and p value
dat2$HR2 = paste0(dat2[, 2], "(", dat2[, 3], "-", dat2[, 4], ")")
dat2$p = ifelse(
  dat2$p <= 0.001,
  "<0.001",
  ifelse(
    dat2$p <= 0.01,
    "<0.01",
    ifelse(
      dat2$p <= 0.05,
      paste(round(dat2$p, 3), "*"),
      round(dat2$p, 3)
    )
  )
)


# plot
labeltext=cbind(c("Factor",dat2$Trait),
                c("HR (95%CI)",dat2$HR2),
                c("P Value",dat2$p))
library(forestplot)
forestplot(
  labeltext,
  mean = c(NA,dat2[, 2]),
  lower = c(NA,dat2[, 3]),
  upper = c(NA,dat2[, 4]),
  zero = 1,
  boxsize = 0.2,
  title="Hazard Ratio Plot (Multivariate Cox)",
  xlab="Odds Ratio",
  is.summary=c(TRUE,rep(FALSE,5)),
  col = fpColors(box = '#1075BB', lines = 'black', zero = 'grey'),
  lty.ci = "solid",
  graph.pos = 3, # position of box plot
  txt_gp=fpTxtGp(label = gpar(cex=1.1),
                 ticks = gpar(cex=1.1),
                 xlab = gpar(cex=1.1),
                 title = gpar(cex=1.2))
)


```


# 2.Nomogram plot and Calibration plot

```{r}
library(rms)
df=surv[,c("event","time","age","N","T","grade","riskscore")]
dd<-datadist(df)
options(datadist="dd")
mod <- cph(as.formula(paste("Surv(time, event) ~ ",paste(colnames(df)[3:7],collapse = "+"))),
           data=df,x=T,y=T,surv = T)

surv<-Survival(mod) 
m1<-function(x) surv(365,x)
m3<-function(x) surv(1095,x)
m5<-function(x) surv(1825,x)

x<-nomogram(mod,
            fun = list(m1,m3,m5),
            funlabel = c('1-y survival',
                         '3-y survival',
                         '5-y survival'),
            lp = F)

plot(x)

### Calibration plot ###

f1 <- cph(formula =  as.formula(paste("Surv(time, event) ~ ",paste(colnames(df)[3:7],collapse = "+"))),
          data=df,x=T,y=T,surv = T, time.inc=365)
cal1 <- calibrate(f1, cmethod="KM", method="boot", u=365, m=50, B=1000)

f3 <- cph(formula =  as.formula(paste("Surv(time, event) ~ ",paste(colnames(df)[3:7],collapse = "+"))),
          data=df,x=T,y=T,surv = T, time.inc=1095)
cal3 <- calibrate(f3, cmethod="KM", method="boot", u=1095, m=50, B=1000)

f5 <- cph(formula =  as.formula(paste("Surv(time, event) ~ ",paste(colnames(df)[3:7],collapse = "+"))),
          data=df,x=T,y=T,surv = T,  time.inc=1825)
cal5 <- calibrate(f5, cmethod="KM", method="boot", u=1825, m=50, B=1000)

# plot
plot(cal1,lwd = 2,lty = 0,errbar.col = "#92C5DE",
     bty = "l", 
     xlim = c(0,1),ylim= c(0,1),
     xlab = "Nomogram-prediced OS (%)",ylab = "Observed OS (%)",
     col = "#92C5DE",
     cex.lab=1.2,cex.axis=1, cex.main=1.2, cex.sub=0.6)
lines(cal1[,c('mean.predicted',"KM")],
      type = 'b', lwd = 2, col = "#92C5DE", pch = 16)
mtext("")

plot(cal3,lwd = 2,lty = 0,errbar.col = "#F4A582",add = T,
     bty = "l",
     xlim = c(0,1),ylim= c(0,1),
     xlab = "Nomogram-prediced OS (%)",ylab = "Observed OS (%)",
     col = "#F4A582",
     cex.lab=1.2,cex.axis=1, cex.main=1.2, cex.sub=0.6)
lines(cal3[,c('mean.predicted',"KM")],
      type = 'b', lwd = 2, col = "#F4A582", pch = 16)
mtext("")

plot(cal5,lwd = 2,lty = 0,errbar.col = "#66C2A5",
     xlim = c(0,1),ylim= c(0,1),col = "#66C2A5",add = T)
lines(cal5[,c('mean.predicted',"KM")],
      type = 'b', lwd = 2, col = "#66C2A5", pch = 16)

abline(0,1, lwd = 2, lty = 3, col = "#224444")

legend("bottomright", 
       legend = c("1-y survival","3-y survival","5-y survival"), 
       col = c("#92C5DE", "#F4A582", "#66C2A5"), 
       lwd = 2,
       cex = 1.2)
```

# 3.DCA curve
```{r}
#devtools::install_github('yikeshu0611/ggDCA')
df$nomo = predict(model1,newdata = df)

library(ggDCA)

nomo  <- cph(Surv(time,event)~nomo,df)
age   <- cph(Surv(time,event)~age,df)
N <- cph(Surv(time,event)~N,df)
T <- cph(Surv(time,event)~T,df)
grade <- cph(Surv(time,event)~grade,df)
dat3  <- dca(nomo,age,N,T,grade)
ggplot(dat3)
```
